backend/requirements.txt
numpy
Flask==2.2.2
Flask-Cors==3.0.10
pillow
Werkzeug==2.2.2
gunicorn==20.1.0